pyproject.toml
[project]
name = "shorteo"
version = "0.1.0"
description = "Service for shortening URLs"
readme = "README.md"
authors = [
    { name = "Jan Okruta", email = "jan.okruta@gmail.com" }
]
requires-python = "==3.13.*"
dependencies = [
    "django~=5.1",
    "django-rest-framework",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    "ipython",
    "mypy",
    "parameterized",
    "pre-commit",
    "ruff",
]

[tool.ruff]
line-length = 96
target-version = "py313"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.mypy]
exclude = [
    "manage.py",
    "test.*\\.py$",
    ".*_test\\.py$",
    "tests/",
    "test/",
]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
python_version = "3.13"
strict = true

[tool.django-stubs]
django_settings_module = "shorteo.settings"
